Updated IntraDayDataProcessor to generate HTML report that lists graphs of each symbol. The Blue line on top is the close price. The Black line under it is the volume. The redline indicates zero for volume. If black line is above the redline, it indicates buy volume, otherwise sell volume. If black line is right over the redline (zero mark), it indicates equal buying and selling.
Any volume that corresponds with a higher price is considered buy and likewise any volume that corresponds with a lower price is considered sell.

Conclusion: Consider buying symbols that have less or no short volume, when there is a spike in buy volume say at least 4 times normal buy volume.